fixin bugs at function run

+++ src/primitive_db/engine.py
@@ -24,16 +24,24 @@ def run():
     
     while True:
         user_input = input("Введите команду: ").strip()
+        if not user_input:
+            print("Введите команду")
+            continue
+
         args = shlex.split(user_input)
         command = args[0].lower()
 
         if command == "exit": #выход
             print("Выход из программы...")
             break
+
         elif command == "help": #список доступных команд
             show_help()
+    
         elif command == "create_table": #создать таблицу
             metadata = load_metadata(FILE_PATH)
+            if metadata is None:
+                metadata = {}
             if len(args) < 3:
                     print("Ошибка: недостаточно аргументов")
                     show_help()
@@ -41,21 +49,35 @@ def run():
             table_name = args[1]
             columns = args[2:] 
             new_metadata = create_table(metadata, table_name, columns)
-            save_metadata(FILE_PATH, new_metadata)
-            print(f"Таблица '{table_name}' успешно создана")
+            if new_metadata is not None:
+                save_metadata(FILE_PATH, new_metadata)
+                print(f"Таблица '{table_name}' успешно создана")
+
         elif command == "list_tables": #список таблиц
             metadata = load_metadata(FILE_PATH)
+            if metadata is None:
+                metadata = {}
             tables = list_tables(metadata)
             if tables is None:
                 print("Пока нет существующих таблиц... Создайте новую таблицу")
-            print(", ".join(tables))
+            else:
+                print(", ".join(tables))
         elif command == "drop_table": #удалить таблицу
+
             metadata = load_metadata(FILE_PATH)
+            if metadata is None:
+                metadata = {}
             table_name = args[1]
+            if len(args) < 2:
+                print("Ошибка: укажите имя таблицы")
+                continue
             new_metadata = drop_table(metadata, table_name)
             save_metadata(FILE_PATH, new_metadata)
+            print(f"Таблица '{table_name}' успешно удалена")
+
         elif command: #неизвестная команда
             print(f"Неизвестная команда: {command}")
             show_help()
+
         else: #нет команды
             print("Введите команду")
